exmples/redis-om-python-app/requirements.txt
redis-om>=0.2.1
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0